Update token amount (2 decimals if needed)

+++ IPSX/Model/UserInfo.swift
@@ -35,7 +35,7 @@ struct UserInfo {
     var countryID: String
     var email: String
     var proxyTest: String
-    var balance: Int
+    var balance: Double
     var kycStatus: String = "Unknown Status Text".localized
     var socialName: String?
     var refferalCode: String?
@@ -43,7 +43,7 @@ struct UserInfo {
     
     //TODO (some new minion in the future): Refactor the constructor to accept a dictionary
     init(firstName: String = "", middleName: String = "",lastName: String = "",
-         telegram: String = "", countryID: String = "", email: String = "", proxyTest: String = "", balance: Int = 0, kycStatus: Int = -1, socialName: String? = nil, refferalCode: String? = nil, deleteAccountDate: Date? = nil, pendingDeleteAccount: Bool = false) {
+         telegram: String = "", countryID: String = "", email: String = "", proxyTest: String = "", balance: Double = 0, kycStatus: Int = -1, socialName: String? = nil, refferalCode: String? = nil, deleteAccountDate: Date? = nil, pendingDeleteAccount: Bool = false) {
         
         self.firstName    = firstName
         self.middleName   = middleName
